requirements.txt
openai>=1.10.0
anthropic>=0.18.0
pypdf>=3.17.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

